server/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pythonocc-core==7.7.0 